[ADD] sale_packaging_default: default packaging for sales
This module simplifies and emphasizes the usage of packaging in sales orders.

- Packaging fields in sale order lines appear before product quantity fields.
- When selecting a product to sell, its default packaging is added automatically.

Why is this module useful? Packaging usage in Odoo is designed to be computed
automatically when you choose the final amount of products to sell. Although
you can still do that, with this module the natural usage is inverted: it's
easier for you to select the packaging and quantity first, and then the final
amount of products is computed automatically.

@moduon MT-3695