[FIX] sale_mrp_bom: accept BoMs without variant
This commit fixes a bug where the module would not accept a BoM on a
Sale Order Line if the product in question had multiple variants and
the BoM did not specify a variant.

Now, regardless of whether the product has multiple variants or not,
if the variant is specified on the BoM, it must match the
product on the Sale Order Line (like before), but if it's
not specified, matching on the product template is enough.