[FIX] sale_order_product_picker: Fix addon installation when this addon is installed
It's a bad idea to use the registry at initialization phase. When you access the registry some info are set into the cache. For exemple the field's ids list of a given model. That means that new field added to the model later into the installation process will not be returned by a new call to the _get_ids method of the ir.model.fields model. At the end, the installation of new addon adding, for exemple, a selection field to the sale.order model is no more possible.