[FIX] sale_blanket_order: filter default taxes by company for all users
**Before this commit**

When adding a product on a sale blanket order line, all the taxes of the product from all accessible companies were added to the blanket order line.

**After this commit**

When adding a product on a sale blanket order line, only the taxes of the product from the company on the blanket order are added to the blanket order line. This mimics the treatment of taxes on sale order lines when adding a new product.

Also, some field and view attributes were updated in analogy to sale.order.line.