[IMP] sale_exception: Confirm directly the sale order when ignoring exceptions
When there are exceptions blocking the confirmation of an order, the user
normally finds out when clicking on the "Confirm" button.

At that time, the wizard is popup and the user -if he has enough permissions-
can ignore the exceptions and click confirm.

Before this commit, the wizard would only flag the order to ignore the
exceptions, but the order would not be confirmed. The user would have to click
again on the "Confirm" button to do so.

This is counter-intuitive since the user intention was clearly to confirm the
order in the first place, and it just forces the user to click again.

After this commit, the order is confirmed directly.

Additionally, the wizard will now hide the "Confirm" button if the exceptions
haven't been ignored.